backend/Provider/requirementsProveedores.txt
fastapi
uvicorn[standard]
pytest
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pydantic==2.8.2
pytest==8.2.2
httpx
fastapi-cache2==0.2.1